Preserve strays in cap-recovery, delete old recorded endpoint by ID, split tests
Address five unresolved review threads on #1827:

1. 'Preserve working strays until cap retry succeeds' (Codex):
   The endpoint-limit fallback now deletes the OLD RECORDED endpoint by
   ID (not same-URL strays) to free a slot. Same-URL strays are preserved
   — if the retry also fails, one of them may still match the DB's stored
   secret, so webhooks keep delivering. The main cleanup (after DB save)
   deletes remaining strays once the new endpoint is durably recorded.

2. 'Delete the previous recorded endpoint after domain changes' (Codex):
   settings-stripe.ts now captures the previous endpoint ID before the
   DB save and passes it to cleanupOldWebhookEndpoints as an explicit
   'also delete' ID. After a domain change the old recorded endpoint is
   at a different URL and won't appear in the same-URL listing, so it
   would never be deleted and would keep consuming Stripe quota.

3. 'Free the old recorded endpoint when cap blocks domain moves' (Codex):
   Fixed by thread 1's approach — the limit-error fallback deletes the
   old recorded endpoint by ID regardless of its URL, freeing the slot
   it was consuming even after a domain move.

4. 'Split the oversized webhook setup test' (Codex):
   test/lib/stripe/webhook-setup.test.ts (478 lines) is split into:
   - webhook-mocks.ts (183 lines) — shared mock helpers
   - webhook-setup.test.ts (271 lines) — setup + client config tests
   - webhook-cleanup.test.ts (113 lines) — cleanup tests
   All files now under the ~400-line target.

5. 'Add a webhook-only error case for isEndpointLimitError' (CodeRabbit):
   New test 're-throws when error is webhook but not limit or maximum'
   uses createThrowsWebhookOnly option returning a 400 with message
   'Invalid webhook URL format' — contains 'webhook' but neither 'limit'
   nor 'maximum', exercising the false branch of isEndpointLimitError.

cleanupOldWebhookEndpoints gains an alsoDeleteIds parameter for the
domain-move case. A new test 'also deletes explicit IDs for domain-move
cleanup' verifies both same-URL strays and the old recorded endpoint
(at a different URL) are deleted.
